feat: create pagination style
