+++ src/components/index.ts
@@ -2,6 +2,7 @@ export { Header } from './organisms/Header';
 export { SearchBar } from './organisms/SearchBar';
 export { Card } from './organisms/Card';
 export { Footer } from './organisms/Footer';
+export { Pagination } from './organisms/Pagination';
 
 export { Type } from './atoms/Type';
 export { Button } from './atoms/Button';
